requirements.txt
boto3==1.11.5
botocore==1.14.5
certifi==2019.11.28
chardet==3.0.4
dj-database-url==0.5.0
Django==2.2.13
django-filter==2.2.0
django-forms-bootstrap==3.1.0
django-storages==1.8
docutils==0.15.2
gunicorn==20.0.4
idna==2.8
jmespath==0.9.4
psycopg2-binary==2.8.4
python-dateutil==2.8.1
pytz==2019.3
requests==2.22.0
s3transfer==0.3.1
six==1.14.0
sqlparse==0.3.0
stripe==2.42.0
urllib3==1.25.7
